Split base functionality into separate RtmApiBase
