+++ rtm-scala/src/main/scala/com/github/vergenzt/rtmscala/RtmApi.scala
@@ -23,80 +23,7 @@ import util.XmlConversions._
  * @define perms @param perms The requested permission.
  * @define timeline @param timeline A timeline obtained from `rtm.timelines.create`
  */
-object RtmApi extends RtmApiGenerated {
-
-  val BASE_URL = "http://api.rememberthemilk.com/services"
-  val AUTH_URL = BASE_URL + "/auth/"
-
-  // for testing
-  private[rtmscala] var REST_URL = BASE_URL + "/rest/"
-
-  /**
-   * Construct an RTM API request.
-   * @param url The url to request.
-   * $params
-   * $creds
-   */
-  protected[rtmscala]
-  def baseRequest(url: String, params: (String, String)*)
-      (implicit creds: ApiCreds): HttpRequest = {
-    Http(url).param("api_key", creds.apiKey).params(params)
-  }
-
-  /**
-   * Construct an unsigned RTM API request.
-   * $method
-   * $params
-   * $creds
-   */
-  protected[rtmscala]
-  def unsignedRequest(method: String, params: (String, String)*)
-      (implicit creds: ApiCreds): HttpRequest = {
-    baseRequest(REST_URL, ("method" -> ("rtm." + method)) :: params.toList: _*)
-  }
-
-  /**
-   * Construct a signed RTM API request.
-   * $method
-   * $params
-   * $creds
-   */
-  protected[rtmscala]
-  def request(method: String, params: (String, String)*)
-      (implicit creds: ApiCreds): HttpRequest = {
-    unsignedRequest(method, params: _*).signed
-  }
-
-  /**
-   * Construct an authenticated RTM API request.
-   * $method
-   * $params
-   * $creds
-   * $authToken
-   */
-  protected[rtmscala]
-  def authedRequest(method: String, params: (String, String)*)
-      (implicit creds: ApiCreds, authToken: AuthToken) = {
-    request(method, ("auth_token" -> authToken.token) :: params.toList: _*)
-  }
-
-  /**
-   * Construct an authenticated RTM API request with a timeline.
-   * $method
-   * $params
-   * $creds
-   * $authToken
-   * $timeline
-   */
-  protected[rtmscala]
-  def timelinedRequest(method: String, params: (String, String)*)
-      (implicit creds: ApiCreds, authToken: AuthToken, timeline: Timeline) = {
-    authedRequest(method, ("timeline" -> timeline.id) :: params.toList: _*)
-  }
-
-  /*****************************
-   * Begin API implementation. *
-   *****************************/
+object RtmApi extends RtmApiBase with RtmApiGenerated {
 
   object auth {
     /**
